refactor: resolveSelectedPattern 유틸로 분리 및 전역 상태 기본값 적용
