+++ src/domain/shuttle/hooks/use-selected-shuttle-pattern-store.ts
@@ -1,15 +1,17 @@
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
 
+export const DEFAULT_PATTERN_CODE = "main_to_jungwang";
+
 interface SelectedShuttlePatternState {
-  readonly selectedPatternCode: string | null;
+  readonly selectedPatternCode: string;
   readonly setSelectedPatternCode: (code: string) => void;
 }
 
 export const useSelectedShuttlePatternStore = create<SelectedShuttlePatternState>()(
   persist(
     (set) => ({
-      selectedPatternCode: null,
+      selectedPatternCode: DEFAULT_PATTERN_CODE,
       setSelectedPatternCode: (code: string): void => {
         set({ selectedPatternCode: code });
       },
